requirements.txt
langchain
langchain-core
langchain-community
langchain-neo4j
langchain-openai
openai
litellm
fastapi
uvicorn
motor
pymongo
nltk
neo4j
llama-index-embeddings-azure-openai
llama-index-llms-azure-openai
langfuse
asyncio
llama-index-vector-stores-redis
llama-index
llama-index-core
python-dotenv
sentence-transformers
redis
beautifulsoup4
scrapy
requests
pydantic
# serper